Generalize classes and interfaces
